task-harmony-email-extraction/requirements.txt
groq
python-dotenv
pydantic
tqdm